// src/interfaces/CarInterface.ts
export interface Cars {
  id: number;
  plate: string;
  manufacture: string;
  model: string;
  image: string;
  image_public_id: string;
  rent_per_day: number;
  capacity: number;
  description: string;
  available_at: Date;
  transmission: string;
  available: boolean;
  type: string;
  year: number;
  options: string;
  specs: string;
  created_at: string;
  updated_at: string;
  deleted_at: string;
}